+++ lihil/oas/__init__.py
@@ -8,7 +8,7 @@ from typing import Any
 
 from lihil.config.app_config import IOASConfig
 from lihil.interface.problem import DetailBase
-from lihil.routing import EndpointProps, Route, RouteBase
+from lihil.routing import Route, RouteBase
 from lihil.utils.json import encoder_factory
 from lihil.vendors import Response
 
@@ -29,7 +29,7 @@ def get_openapi_route(
             oas = generate_oas(routes, oas_config, app_version)
         return Response(encoder(oas), media_type="application/json")
 
-    openapi_route = Route(oas_config.OAS_PATH, props=EndpointProps(in_schema=False))
+    openapi_route = Route(oas_config.OAS_PATH, in_schema=False)
     openapi_route.get(openapi)
     return openapi_route
 
@@ -41,7 +41,7 @@ def get_doc_route(oas_config: IOASConfig) -> Route:
     async def swagger():
         return get_swagger_ui_html(openapi_url=oas_path, title=oas_config.TITLE)
 
-    doc_route = Route(docs_path, props=EndpointProps(in_schema=False))
+    doc_route = Route(docs_path, in_schema=False)
     doc_route.get(swagger)
     return doc_route
 
@@ -54,6 +54,6 @@ def get_problem_route(
     async def problem_detail():
         return get_problem_ui_html(title=oas_config.PROBLEM_TITLE, problems=problems)
 
-    problem_route = Route(problem_path, props=EndpointProps(in_schema=False))
+    problem_route = Route(problem_path, in_schema=False)
     problem_route.get(problem_detail)
     return problem_route